Web/Backend/requirements.txt
aiohttp
cachetools
ecdsa
eth-hash[pycryptodome]
execnet
fastecdsa
frozendict
gprof2dot
lark
marshmallow-dataclass>=7.1.0
marshmallow-enum
marshmallow-oneofschema
marshmallow>=3.2.1
mpmath
numpy
pipdeptree
prometheus-client
pytest
pytest-asyncio
pytest-profiling
pytest-xdist
PyYAML
sympy
typeguard<3.0.0
web3
flask
flask_sqlalchemy
flask_cors
flask_migrate
torch
transformers
huggingface-hub
requests
cairosvg
cairo-lang
cairocffi
cssselect2
defusedxml
tinycss2
Pillow
sentence_transformers
chromadb
torchvision
SentencePiece